+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Add the backup memory type file generation for Mednafen
 * Rewrite interface into python scripts instead of glade files
 * Fix game with custom name won't work correctly in self.__on_game_launch
+* Fix shortcuts in preferences are not store correctly in main configuration
 
 ## Version 0.6.1 - Dr. Funfrock (30 Mar, 2017)